remove separate SenderAgentMessageId

+++ src/Simplex/Messaging/Agent/Store.hs
@@ -150,7 +150,7 @@ data RcvMsg = RcvMsg
 
 type InternalRcvId = Int64
 
-type ExternalSndId = Integer
+type ExternalSndId = Int64
 
 type ExternalSndTs = UTCTime
